Publish master to opensource
commit d0e2a0f27a88221440776a4b737e421bef32edfe
Merge: 5d6eef8 2559529
Date:   Fri Dec 4 13:25:04 2015 -0500

    Merge pull request #48 from ameredit/feature/fix_warnings_in_headers

    Feature/fix warnings in headers

commit 255952990536c7442954767cfbe50c623608aa77
Merge: dd2241d e9fe3a0
Date:   Tue Nov 24 17:35:24 2015 -0500

    Merge bbgithub.dev.bloomberg.com:bde/bde into feature/fix_warnings_in_headers

commit dd2241d49804d11ae8ece3d99c2f0fa514596b07
Date:   Tue Nov 24 17:35:03 2015 -0500

    Resolve narrowing warning in bslst_stringref header

    Strictly speaking, this patch silences a warning that is giving
    us useful information on a 64-bit platform, when trafficking in
    references to strings longer than 2-billion elements.  In practice,
    that is not something we will need to worry about for a very long
    time, so we silence the warning in favor of existing practice.

commit bd267445b029c432cb156ea00c1d62ff2f465076
Date:   Tue Nov 24 17:31:36 2015 -0500

    Fix byte-count data type to match allocators

    'int' is no longer a suitable type to track memory usagae, in
    a 64-bit world.  Use the 'size_type' from the underlying
    allocator, which maps to 'std::size_t' and will have the
    appropiate size on all platforms, although also switches from
    signed to unsigned.  This resolves a number of warnings
    originating in a header file.

commit 4ff002bf7c7e029db037335e4dfccdcfd9a628fe
Date:   Tue Nov 24 17:28:58 2015 -0500

    Fix warning about redundant 'const'

    When fundamental types are returned by value, the resulting rvalue
    is immutable, effectively 'const', whether the keyword is used or
    not.  Compilers like to warn about this.

commit dbc2a387cbc77e894c4933605c06fa1f35f723de
Date:   Tue Nov 24 17:27:22 2015 -0500

    Make StringRefs directly from 'string's

    Rather than construct a 'StringRef' from calling two methods
    on a 'bslstl::string' object, and incurring a warning, simply
    pass the string to the constructor of 'bslstl::StringRef',
    whilch will do the correct thing, most efficiently, and without
    a warning.

commit 7664341d99843232401c58558cfa9c0380e3155c
Date:   Tue Nov 24 14:34:46 2015 -0500

    Fix warning in ball_userfieldsschama.h

    Warning on converting a 'size_t' to an 'int', which could
    be a significant narrowing on a 64-bit platform.  That said,
    2 billion field values per object  should be more than enough for
    much longer than the immediate future, so the easiest fix is
    to cast to 'int', rather than change the return type.

    Considered adding a 'BSLS_ASSERT_SAFE' that the value was
    less than 'numeric_limits<int>::max()', but did not want to
    add an extra header dependency in this header.  Would have
    done so if the method were out-of-line in the .cpp file.

commit bd8c415f4b88492d18d600f305e415309671fd1f
Date:   Tue Nov 24 14:32:13 2015 -0500

    Fix warning in ball_userfieldsschama.h

    Warning on converting a 'size_t' to an 'int', which could
    be a significant narrowing on a 64-bit platform.  That said,
    2 billion fields in a schema should be more than enough for
    much longer than the immediate future, so the easiest fix is
    to cast to 'int', rather than change the return type.

    Considered adding a 'BSLS_ASSERT_SAFE' that the value was
    less than 'numeric_limits<int>::max()', but did not want to
    add an extra header dependency in this header.  Would have
    done so if the method were out-of-line in the .cpp file.

+++ groups/bsl/bslstl/bslstl_stringref.h
@@ -953,7 +953,8 @@ int StringRefImp<CHAR_TYPE>::compare(
                     other.data(),
                     native_std::min(this->length(), other.length()));
 
-    return result != 0 ? result : this->length() - other.length();
+    return result != 0 ? result
+                       : static_cast<int>(this->length() - other.length());
 }
 
 }  // close package namespace
